LayoutTests/fast/hidpi/image-srcset-fraction-expected.txt

blue-100-px-square.png has MIME type image/png
image-srcset-fraction.html has MIME type text/html
srcset-helper.js has MIME type text/javascript
js-test-pre.js has MIME type text/javascript
green-200-px-square.png has MIME type image/png
PASS document.getElementById("testimg").clientWidth==200 is true
This test passes if the srcset resource is loaded and displayed as the image
